requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
django-cors-headers>=3.13.0, <3.14
django-storages>=1.13.1, <1.14
boto3>=1.26.33, <1.27
Pillow>=9.3.0, <9.4
requests>=2.28.1, <2.29
ebaysdk>=2.2.0, <2.3